refactor: update iOS app to use new engine framework

+++ Wolfenstein3d-Swift-CPU/Managers/InputManager.swift
@@ -1,25 +1,20 @@
-// Managers/InputManager.swift
 import UIKit
+import Engine
 
 class InputManager: InputProviderProtocol {
-  // MARK: - Properties
-  
   weak var delegate: InputHandlerProtocol?
   
   private var moveX: CGFloat = 0
   private var moveY: CGFloat = 0
   private var isMoving = false
   
-  // MARK: - Public Methods
-  
   func processJoystickInput(x: CGFloat, y: CGFloat) {
     moveX = x
     moveY = y
-    isMoving = abs(x) > GameConfig.Input.joystickDeadzone ||
-    abs(y) > GameConfig.Input.joystickDeadzone
+    isMoving = abs(x) > CGFloat(GameConfig.Input.joystickDeadzone) ||
+    abs(y) > CGFloat(GameConfig.Input.joystickDeadzone)
     
     if isMoving {
-      // Y is inverted: joystick up = negative Y, but forward = positive
       delegate?.handleMovement(forward: Double(-moveY), strafe: Double(moveX))
     }
   }
@@ -28,7 +23,6 @@ class InputManager: InputProviderProtocol {
     moveX = 0
     moveY = 0
     isMoving = false
-    // Reset movement when joystick is released
     delegate?.handleMovement(forward: 0, strafe: 0)
   }
   
@@ -39,17 +33,11 @@ class InputManager: InputProviderProtocol {
   }
   
   func processTouchLook(deltaX: CGFloat, deltaY: CGFloat) {
-    // Convert touch delta to rotation
-    // Horizontal movement rotates the player
     if abs(deltaX) > 0 {
       delegate?.handleRotation(angle: Double(deltaX))
     }
     
-    // Note: deltaY could be used for vertical look if implementing full 3D look
-    // For now, Wolfenstein 3D style only uses horizontal rotation
   }
-  
-  // MARK: - State Queries
   
   func isPlayerMoving() -> Bool {
     return isMoving
